Render AnnouncementShowLive from AnnouncementController#show
What changed?
===========

We render the AnnouncementShowLive instead of rendering the static page
for the announcement. This will allow us to move the commenting
completely inside the live view along with the subscribe/unsubscribe
from the announcement

Instead of rendering from the router directly, we render it from the
controller. Typically, I'd go directly from the router, but this has two
benefits:

- we already have the current user, so we pass that as session data
- we do deslugifying of the `id` in the controller, so we don't need to
  move that out of the AnnouncmentController