Import Phoenix.HTML functions needed for EmailsTest
The emails test required `link` and `safe_to_string` functions from
`Phoenix.HTML`. Previously, it was using `ConstableWeb, :view` in
`ConnCase` to get those functions, which is really heavy handed and
affected the rest of the tests using `ConnCase`

This simply imports the functions needed in the test since they're
something of a special case.