Update comment only via pubsub
We remove the updating of the comment in memory in the
AnnouncementShowLive when a new comment is created. Since we do it via
PubSub, and since all other live views are receiving the pub sub, the
current user's live view also receives the update -- so the new comment
would show twice if we update it immediately in memory and then once
again via pub sub.