Update NewComment feature test to target a new ID
The NewComment acceptance tests was no longer passing because for some
reason it could no longer find the "comment_body". The `text_field`
Wallaby function takes in a selector which can be an id, name, or
placeholder of the field (or the label). We update it to target the id,
since it seems like the change from `form_for` to `f = form_for` for
LiveView might have changed the input's name.