Move ImageUploader into a LiveView Hook
We move the ImageUploader into a Phoenix LiveView Hook. The nice thing
about this is that we can make it agnostic of the element's id. We
simply set the `ImageUploader` hook on the element and then since it's a
hook, the same functionality applies. So even though we only use it for
comment forms right now, we can use it in the future with announcements.