Separate comment preview from user autocomplete
What changed?
============

We separate the comment preview logic (still in comment-form.js, even
though it's not really a form anymore) from the user autocomplete logic.

In order to render the user autocomplete, we need to have the data
rendered on the page. So that's more difficult to do with LiveView
Hooks. But the rest of the comment preview (as well as saving the
comment in localStorage) can be done in a LiveView hook.

That fixes the issue of the JS loading strangely on the page. The form
now looks back to normal.